requirements.txt
beautifulsoup4==4.13.5
requests==2.32.5
pandas==2.3.2
selenium==4.35.0
lxml==6.0.1
python-dotenv==1.1.1
google-api-python-client==2.182.0